Patch 1149
・mob_talk() 実行するとき、最大値の判定を行うように強化（mob.c）
・pc_break_equip() の引数 where をunsinged short型からint型に拡張し、全てのLOC定数を受け入れることが
  可能にしておく（pc.*）
・map_session_data構造体の unbreakable_equip メンバをunsigned short型からunsinged int型に拡張し、
  全てのLOC定数でビット演算が可能にしておく（map.h）
・clif.cのリファクタリング、武器の見た目を取得する処理を clif_get_weapon_view() としてまとめる（clif.c）
・PACKETVERが3以下のとき、clif_changelook() でコンパイルエラーが発生する問題を修正（clif.c）
・装備箇所を示すLOC系定数をitemdb.hからmmo.hに移動し、char.cでも利用可能にする（mmo.h, itemdb.h）
・出来る限りenum型を使うように変更
  （chardb_txt.c, char.c, clif.c, extra.c, mob.c, pc.c, script.c, skill.c, status.c）

+++ src/map/map.h
@@ -418,7 +418,7 @@ struct map_session_data {
 	unsigned int inventory_sortkey;
 	unsigned int cart_sortkey;
 	short equip_index[EQUIP_INDEX_MAX];
-	unsigned short unbreakable_equip;
+	unsigned int unbreakable_equip;
 	int weight,max_weight;
 	int cart_weight,cart_max_weight,cart_num,cart_max_num;
 	char mapname[24];
@@ -508,7 +508,7 @@ struct map_session_data {
 	short view_class;
 	struct pc_base_job s_class;
 
-	short weapontype1,weapontype2;
+	short weapontype1,weapontype2;	// 〜WT_MAX
 	int paramb[6],paramc[6],parame[6],paramcard[6];
 	int hit,flee,flee2,aspd,amotion,dmotion;
 	int watk,watk2,atkmods[MAX_SIZE_FIX];
